requirements.txt
absl-py==2.1.0
aiomysql==0.2.0
aiopg==1.4.0
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.2.0
autopep8==2.3.1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.16.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
databases==0.9.0
debugpy==1.8.1
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.0.1
fastapi~=0.115.6
fastapi-cli==0.0.5
flatbuffers==24.3.25
fonttools==4.51.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jax==0.4.27
jaxlib==0.4.27
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.12
packaging==24.0
parso==0.8.4
passlib==1.7.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.44
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pydantic~=2.10.3
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.1
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
pytweening==1.2.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
rich==13.9.4
rsa==4.9
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
spyder-kernels==2.5.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
traitlets==5.14.3
typer==0.14.0
typing_extensions==4.12.2
ujson==5.10.0
unicorn==2.1.1
uvicorn==0.32.1
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1

jwt~=1.3.1
requests~=2.32.3
pip~=24.3.1
base~=0.0.0
Faker~=33.1.0